cloud_functions/requirements.txt
google-cloud-bigquery[pandas]
db-dtypes
scikit-learn
numpy
scipy
pandas
pytz==2023.3.post1